requirements.txt

fastapi 
sse-starlette 
msgpack_numpy
substrate-interface
paramiko
torch
streamlit
pytest
munch
nest_asyncio
asyncio
aiofiles
retry
netaddr

#
loguru
plotly
openai
aiohttp
communex


